Add ponyfill for injectIntl and intlShape for react-intl>=3
This commit updates the usages of `injectIntl` and `intlShape` to use
the ponyfill inside `./intl/utils.jsx`, so we can use this package with
apps using the newer versions of intl. I also created a jscodeshift
codemod to update the imports of these two functions to the local file
and added it here for posterity. To run the codemod, you can use the
following command in the root of the repository:

jscodeshift --ignore-config=.gitignore -t jscodeshift/upgrade-react-intl.js react