Add tests for GeolocationInput component
Unfortunately, the GeolocationInput component depends heavily on
function defined by the Google Maps SDK. I found the
`@googlemaps/jest-mocks` lib, which does a pretty good job mocking all
methods the SDK exposes, including the event system exposed in
`google.maps.event`.

This raises an issue which we can't actually use this event system to
emit new events to the component (e.g. to emulate selecting an
autocomplete option). And so we need to use the mock calls to
`google.maps.event.addListener` to get the listener callback, and
manually invoke that instead. It would be nice to eventually have at
least this event system working to have tests that are more resilient.